Refactor new shared Publication model, merge Publication and WebPublication
