+++ r2-shared/src/main/java/org/readium/r2/shared/publication/metadata/Metadata.kt
@@ -1,0 +1,191 @@
+/*
+ * Module: r2-shared-kotlin
+ * Developers: Aferdita Muriqi, Clément Baumann, Mickaël Menu
+ *
+ * Copyright (c) 2020. Readium Foundation. All rights reserved.
+ * Use of this source code is governed by a BSD-style license which is detailed in the
+ * LICENSE file present in the project repository where this source code is maintained.
+ */
+
+package org.readium.r2.shared.publication.metadata
+
+import org.json.JSONObject
+import org.readium.r2.shared.JSONable
+import org.readium.r2.shared.Warning
+import org.readium.r2.shared.WarningLogger
+import org.readium.r2.shared.extensions.*
+import org.readium.r2.shared.extensions.putIfNotEmpty
+import org.readium.r2.shared.publication.LocalizedString
+import org.readium.r2.shared.publication.ReadingProgression
+import org.readium.r2.shared.publication.link.LinkHrefNormalizer
+import org.readium.r2.shared.publication.link.LinkHrefNormalizerIdentity
+import java.io.Serializable
+import java.util.*
+
+/**
+ * https://readium.org/webpub-manifest/schema/metadata.schema.json
+ *
+ * @param readingProgression WARNING: This contains the reading progression as declared in the
+ *     publication, so it might be [AUTO]. To lay out the content, use
+ *     [publication.contentLayout.readingProgression] to get the calculated reading progression from
+ *     the declared direction and the language.
+ * @param otherMetadata Additional metadata for extensions, as a JSON dictionary.
+ */
+data class Metadata(
+    val identifier: String? = null, // URI
+    val type: String? = null, // URI (@type)
+    val localizedTitle: LocalizedString,
+    val localizedSubtitle: LocalizedString? = null,
+    val modified: Date? = null,
+    val published: Date? = null,
+    val languages: List<String> = emptyList(), // BCP 47 tag
+    val sortAs: String? = null,
+    val subjects: List<Subject> = emptyList(),
+    val authors: List<Contributor> = emptyList(),
+    val translators: List<Contributor> = emptyList(),
+    val editors: List<Contributor> = emptyList(),
+    val artists: List<Contributor> = emptyList(),
+    val illustrators: List<Contributor> = emptyList(),
+    val letterers: List<Contributor> = emptyList(),
+    val pencilers: List<Contributor> = emptyList(),
+    val colorists: List<Contributor> = emptyList(),
+    val inkers: List<Contributor> = emptyList(),
+    val narrators: List<Contributor> = emptyList(),
+    val contributors: List<Contributor> = emptyList(),
+    val publishers: List<Contributor> = emptyList(),
+    val imprints: List<Contributor> = emptyList(),
+    val readingProgression: ReadingProgression = ReadingProgression.AUTO,
+    val description: String? = null,
+    val duration: Double? = null,
+    val numberOfPages: Int? = null,
+    val belongsToCollections: List<Collection> = emptyList(),
+    val belongsToSeries: List<Collection> = emptyList(),
+    val otherMetadata: Map<String, Any> = mapOf()
+) : JSONable, Serializable {
+
+    /**
+     * Serializes a [Metadata] to its RWPM JSON representation.
+     */
+    override fun toJSON() = JSONObject(otherMetadata).apply {
+        put("identifier", identifier)
+        put("@type", type)
+        putIfNotEmpty("title", localizedTitle)
+        putIfNotEmpty("subtitle", localizedSubtitle)
+        put("modified", modified?.toIso8601String())
+        put("published", published?.toIso8601String())
+        putIfNotEmpty("language", languages)
+        put("sortAs", sortAs)
+        putIfNotEmpty("subject", subjects)
+        putIfNotEmpty("author", authors)
+        putIfNotEmpty("translator", translators)
+        putIfNotEmpty("editor", editors)
+        putIfNotEmpty("artist", artists)
+        putIfNotEmpty("illustrator", illustrators)
+        putIfNotEmpty("letterer", letterers)
+        putIfNotEmpty("penciler", pencilers)
+        putIfNotEmpty("colorist", colorists)
+        putIfNotEmpty("inker", inkers)
+        putIfNotEmpty("narrator", narrators)
+        putIfNotEmpty("contributor", contributors)
+        putIfNotEmpty("publisher", publishers)
+        putIfNotEmpty("imprint", imprints)
+        put("readingProgression", readingProgression.value)
+        put("description", description)
+        put("duration", duration)
+        put("numberOfPages", numberOfPages)
+        putIfNotEmpty("belongsTo", JSONObject().apply {
+            putIfNotEmpty("collection", belongsToCollections)
+            putIfNotEmpty("series", belongsToSeries)
+        })
+    }
+
+    /**
+     * Syntactic sugar to access the [otherMetadata] values by subscripting [Metadata] directly.
+     * `metadata["layout"] == metadata.otherMetadata["layout"]`
+     */
+    operator fun get(key: String): Any? = otherMetadata[key]
+
+    companion object {
+
+        /**
+         * Parses a [Metadata] from its RWPM JSON representation.
+         *
+         * If the metadata can't be parsed, a warning will be logged with [warnings].
+         */
+        fun fromJSON(
+            json: JSONObject?,
+            normalizeHref: LinkHrefNormalizer = LinkHrefNormalizerIdentity,
+            warnings: WarningLogger? = null
+        ): Metadata? {
+            json ?: return null
+            val localizedTitle = LocalizedString.fromJSON(json.remove("title"), warnings)
+            if (localizedTitle == null) {
+                warnings?.log(Warning.JsonParsing(Metadata::class.java, "[title] is required", json))
+                return null
+            }
+
+            val identifier = json.remove("identifier") as? String
+            val type = json.remove("@type") as? String
+            val localizedSubtitle = LocalizedString.fromJSON(json.remove("subtitle"), warnings)
+            val modified = (json.remove("modified") as? String)?.toIso8601Date()
+            val published = (json.remove("published") as? String)?.toIso8601Date()
+            val languages = json.optStringsFromArrayOrSingle("language", remove = true)
+            val sortAs = json.remove("sortAs") as? String
+            val subjects = Subject.fromJSONArray(json.remove("subject"), normalizeHref, warnings)
+            val authors = Contributor.fromJSONArray(json.remove("author"), normalizeHref, warnings)
+            val translators = Contributor.fromJSONArray(json.remove("translator"), normalizeHref, warnings)
+            val editors = Contributor.fromJSONArray(json.remove("editor"), normalizeHref, warnings)
+            val artists = Contributor.fromJSONArray(json.remove("artist"), normalizeHref, warnings)
+            val illustrators = Contributor.fromJSONArray(json.remove("illustrator"), normalizeHref, warnings)
+            val letterers = Contributor.fromJSONArray(json.remove("letterer"), normalizeHref, warnings)
+            val pencilers = Contributor.fromJSONArray(json.remove("penciler"), normalizeHref, warnings)
+            val colorists = Contributor.fromJSONArray(json.remove("colorist"), normalizeHref, warnings)
+            val inkers = Contributor.fromJSONArray(json.remove("inker"), normalizeHref, warnings)
+            val narrators = Contributor.fromJSONArray(json.remove("narrator"), normalizeHref, warnings)
+            val contributors = Contributor.fromJSONArray(json.remove("contributor"), normalizeHref, warnings)
+            val publishers = Contributor.fromJSONArray(json.remove("publisher"), normalizeHref, warnings)
+            val imprints = Contributor.fromJSONArray(json.remove("imprint"), normalizeHref, warnings)
+            val readingProgression = ReadingProgression.from(json.remove("readingProgression") as? String)
+            val description = json.remove("description") as? String
+            val duration = json.optPositiveDouble("duration", remove = true)
+            val numberOfPages = json.optPositiveInt("numberOfPages", remove = true)
+            val belongsTo = json.remove("belongsTo") as? JSONObject
+            val belongsToCollections = Collection.fromJSONArray(belongsTo?.opt("collection"), normalizeHref, warnings)
+            val belongsToSeries = Collection.fromJSONArray(belongsTo?.opt("series"), normalizeHref, warnings)
+
+            return Metadata(
+                identifier = identifier,
+                type = type,
+                localizedTitle = localizedTitle,
+                localizedSubtitle = localizedSubtitle,
+                modified = modified,
+                published = published,
+                languages = languages,
+                sortAs = sortAs,
+                subjects = subjects,
+                authors = authors,
+                translators = translators,
+                editors = editors,
+                artists = artists,
+                illustrators = illustrators,
+                letterers = letterers,
+                pencilers = pencilers,
+                colorists = colorists,
+                inkers = inkers,
+                narrators = narrators,
+                contributors = contributors,
+                publishers = publishers,
+                imprints = imprints,
+                readingProgression = readingProgression,
+                description = description,
+                duration = duration,
+                numberOfPages = numberOfPages,
+                belongsToCollections = belongsToCollections,
+                belongsToSeries = belongsToSeries,
+                otherMetadata = json.toMap()
+            )
+        }
+
+    }
+
+}
