--- r2-shared/src/main/java/org/readium/r2/shared/publication/webpub/metadata/Collection.kt
@@ -1,16 +1,0 @@
-/*
- * Module: r2-shared-kotlin
- * Developers: Mickaël Menu
- *
- * Copyright (c) 2020. Readium Foundation. All rights reserved.
- * Use of this source code is governed by a BSD-style license which is detailed in the
- * LICENSE file present in the project repository where this source code is maintained.
- */
-
-package org.readium.r2.shared.publication.webpub.metadata
-
-/**
- * Collection type used for collection/series metadata.
- * For convenience, the JSON schema reuse the Contributor's definition.
- */
-typealias Collection = Contributor
