+++ r2-shared/src/test/java/org/readium/r2/shared/publication/ContentLayoutTest.kt
@@ -11,7 +11,6 @@ package org.readium.r2.shared.publication
 
 import org.junit.Assert.*
 import org.junit.Test
-import org.readium.r2.shared.publication.webpub.ReadingProgression
 
 class ContentLayoutTest {
