+++ r2-shared/src/test/java/org/readium/r2/shared/publication/PublicationTest.kt
@@ -9,13 +9,13 @@
 
 package org.readium.r2.shared.publication
 
+import org.json.JSONObject
 import org.junit.Assert.*
 import org.junit.Test
-import org.readium.r2.shared.publication.webpub.LocalizedString
-import org.readium.r2.shared.publication.webpub.ReadingProgression
-import org.readium.r2.shared.publication.webpub.WebPublication
-import org.readium.r2.shared.publication.webpub.link.Link
-import org.readium.r2.shared.publication.webpub.metadata.Metadata
+import org.readium.r2.shared.assertJSONEquals
+import org.readium.r2.shared.publication.link.Link
+import org.readium.r2.shared.publication.metadata.Metadata
+import java.net.URL
 
 class PublicationTest {
 
@@ -27,17 +27,246 @@ class PublicationTest {
         readingOrder: List<Link> = emptyList(),
         resources: List<Link> = emptyList()
     ) = Publication(
-        webpub = WebPublication(
-            metadata = Metadata(
-                localizedTitle = LocalizedString(title),
-                languages = listOf(language),
-                readingProgression = readingProgression
-            ),
-            links = links,
-            readingOrder = readingOrder,
-            resources = resources
-        )
+        metadata = Metadata(
+            localizedTitle = LocalizedString(title),
+            languages = listOf(language),
+            readingProgression = readingProgression
+        ),
+        links = links,
+        readingOrder = readingOrder,
+        resources = resources
     )
+
+    @Test fun `parse minimal JSON`() {
+        assertEquals(
+            Publication(
+                metadata = Metadata(localizedTitle = LocalizedString("Title")),
+                links = listOf(Link(href = "/manifest.json", rels = listOf("self"))),
+                readingOrder = listOf(Link(href = "/chap1.html", type = "text/html"))
+            ),
+            Publication.fromJSON(JSONObject("""{
+                "metadata": {"title": "Title"},
+                "links": [
+                    {"href": "/manifest.json", "rel": "self"}
+                ],
+                "readingOrder": [
+                    {"href": "/chap1.html", "type": "text/html"}
+                ]
+            }"""))
+        )
+    }
+
+    @Test fun `parse full JSON`() {
+        assertEquals(
+            Publication(
+                context = listOf("https://readium.org/webpub-manifest/context.jsonld"),
+                metadata = Metadata(localizedTitle = LocalizedString("Title")),
+                links = listOf(Link(href = "/manifest.json", rels = listOf("self"))),
+                readingOrder = listOf(Link(href = "/chap1.html", type = "text/html")),
+                resources = listOf(Link(href = "/image.png", type = "image/png")),
+                tableOfContents = listOf(Link(href = "/cover.html"), Link(href = "/chap1.html")),
+                otherCollections = listOf(PublicationCollection(role = "sub", links = listOf(Link(href = "/sublink"))))
+            ),
+            Publication.fromJSON(JSONObject("""{
+                "@context": "https://readium.org/webpub-manifest/context.jsonld",
+                "metadata": {"title": "Title"},
+                "links": [
+                    {"href": "/manifest.json", "rel": "self"}
+                ],
+                "readingOrder": [
+                    {"href": "/chap1.html", "type": "text/html"}
+                ],
+                "resources": [
+                    {"href": "/image.png", "type": "image/png"}
+                ],
+                "toc": [
+                    {"href": "/cover.html"},
+                    {"href": "/chap1.html"}
+                ],
+                "sub": {
+                    "links": [
+                        {"href": "/sublink"}
+                    ]
+                }
+            }"""))
+        )
+    }
+
+    @Test fun `parse JSON {context} as array`() {
+        assertEquals(
+            Publication(
+                context = listOf("context1", "context2"),
+                metadata = Metadata(localizedTitle = LocalizedString("Title")),
+                links = listOf(Link(href = "/manifest.json", rels = listOf("self"))),
+                readingOrder = listOf(Link(href = "/chap1.html", type = "text/html"))
+            ),
+            Publication.fromJSON(JSONObject("""{
+                "@context": ["context1", "context2"],
+                "metadata": {"title": "Title"},
+                "links": [
+                    {"href": "/manifest.json", "rel": "self"}
+                ],
+                "readingOrder": [
+                    {"href": "/chap1.html", "type": "text/html"}
+                ]
+            }"""))
+        )
+    }
+
+    @Test fun `parse JSON requires {metadata}`() {
+        assertNull(Publication.fromJSON(JSONObject("""{
+                "links": [
+                    {"href": "/manifest.json", "rel": "self"}
+                ],
+                "readingOrder": [
+                    {"href": "/chap1.html", "type": "text/html"}
+                ]
+        }""")))
+    }
+
+    // {readingOrder} used to be {spine}, so we parse {spine} as a fallback.
+    @Test fun `parse JSON {spine} as {readingOrder}`() {
+        assertEquals(
+            Publication(
+                metadata = Metadata(localizedTitle = LocalizedString("Title")),
+                links = listOf(Link(href = "/manifest.json", rels = listOf("self"))),
+                readingOrder = listOf(Link(href = "/chap1.html", type = "text/html"))
+            ),
+            Publication.fromJSON(JSONObject("""{
+                "metadata": {"title": "Title"},
+                "links": [
+                    {"href": "/manifest.json", "rel": "self"}
+                ],
+                "spine": [
+                    {"href": "/chap1.html", "type": "text/html"}
+                ]
+            }"""))
+        )
+    }
+
+    @Test fun `parse JSON ignores {links} without {rel}`() {
+        assertEquals(
+            Publication(
+                metadata = Metadata(localizedTitle = LocalizedString("Title")),
+                links = listOf(
+                    Link(href = "/manifest.json", rels = listOf("self")),
+                    Link(href = "/withrel", rels = listOf("withrel"))
+                ),
+                readingOrder = listOf(Link(href = "/chap1.html", type = "text/html"))
+            ),
+            Publication.fromJSON(JSONObject("""{
+                "metadata": {"title": "Title"},
+                "links": [
+                    {"href": "/manifest.json", "rel": "self"},
+                    {"href": "/withrel", "rel": "withrel"},
+                    {"href": "/withoutrel"}
+                ],
+                "readingOrder": [
+                    {"href": "/chap1.html", "type": "text/html"}
+                ]
+            }"""))
+        )
+    }
+
+    @Test fun `parse JSON ignores {readingOrder} without {type}`() {
+        assertEquals(
+            Publication(
+                metadata = Metadata(localizedTitle = LocalizedString("Title")),
+                links = listOf(Link(href = "/manifest.json", rels = listOf("self"))),
+                readingOrder = listOf(Link(href = "/chap1.html", type = "text/html"))
+            ),
+            Publication.fromJSON(JSONObject("""{
+                "metadata": {"title": "Title"},
+                "links": [
+                    {"href": "/manifest.json", "rel": "self"}
+                ],
+                "readingOrder": [
+                    {"href": "/chap1.html", "type": "text/html"},
+                    {"href": "/chap2.html"}
+                ]
+            }"""))
+        )
+    }
+
+    @Test fun `parse JSON ignores {resources} without {type}`() {
+        assertEquals(
+            Publication(
+                metadata = Metadata(localizedTitle = LocalizedString("Title")),
+                links = listOf(Link(href = "/manifest.json", rels = listOf("self"))),
+                readingOrder = listOf(Link(href = "/chap1.html", type = "text/html")),
+                resources = listOf(Link(href = "/withtype", type = "text/html"))
+            ),
+            Publication.fromJSON(JSONObject("""{
+                "metadata": {"title": "Title"},
+                "links": [
+                    {"href": "/manifest.json", "rel": "self"}
+                ],
+                "readingOrder": [
+                    {"href": "/chap1.html", "type": "text/html"},
+                ],
+                "resources": [
+                    {"href": "/withtype", "type": "text/html"},
+                    {"href": "/withouttype"}
+                ]
+            }"""))
+        )
+    }
+
+    @Test fun `get minimal JSON`() {
+        assertJSONEquals(
+            JSONObject("""{
+                "metadata": {"title": {"UND": "Title"}, "readingProgression": "auto"},
+                "links": [
+                    {"href": "/manifest.json", "rel": ["self"], "templated": false}
+                ],
+                "readingOrder": [
+                    {"href": "/chap1.html", "type": "text/html", "templated": false}
+                ]
+            }"""),
+            Publication(
+                metadata = Metadata(localizedTitle = LocalizedString("Title")),
+                links = listOf(Link(href = "/manifest.json", rels = listOf("self"))),
+                readingOrder = listOf(Link(href = "/chap1.html", type = "text/html"))
+            ).toJSON()
+        )
+    }
+
+    @Test fun `get full JSON`() {
+        assertJSONEquals(
+            JSONObject("""{
+                "@context": ["https://readium.org/webpub-manifest/context.jsonld"],
+                "metadata": {"title": {"UND": "Title"}, "readingProgression": "auto"},
+                "links": [
+                    {"href": "/manifest.json", "rel": ["self"], "templated": false}
+                ],
+                "readingOrder": [
+                    {"href": "/chap1.html", "type": "text/html", "templated": false}
+                ],
+                "resources": [
+                    {"href": "/image.png", "type": "image/png", "templated": false}
+                ],
+                "toc": [
+                    {"href": "/cover.html", "templated": false},
+                    {"href": "/chap1.html", "templated": false}
+                ],
+                "sub": {
+                    "metadata": {},
+                    "links": [
+                        {"href": "/sublink", "templated": false}
+                    ]
+                }
+            }"""),
+            Publication(
+                context = listOf("https://readium.org/webpub-manifest/context.jsonld"),
+                metadata = Metadata(localizedTitle = LocalizedString("Title")),
+                links = listOf(Link(href = "/manifest.json", rels = listOf("self"))),
+                readingOrder = listOf(Link(href = "/chap1.html", type = "text/html")),
+                resources = listOf(Link(href = "/image.png", type = "image/png")),
+                tableOfContents = listOf(Link(href = "/cover.html"), Link(href = "/chap1.html")),
+                otherCollections = listOf(PublicationCollection(role = "sub", links = listOf(Link(href = "/sublink"))))
+            ).toJSON()
+        )
+    }
 
     @Test fun `get {contentLayout} for the default language`() {
         assertEquals(
@@ -61,6 +290,42 @@ class PublicationTest {
                 readingProgression = ReadingProgression.RTL
             ).contentLayoutForLanguage("EN")
         )
+    }
+
+    @Test fun `set {self} link`() {
+        val publication = createPublication()
+        publication.setSelfLink("http://manifest.json")
+
+        assertEquals(
+            "http://manifest.json",
+            publication.linkWithRel("self")?.href
+        )
+    }
+
+    @Test fun `set {self} link replaces existing {self} link`() {
+        val publication = createPublication(
+            links = listOf(Link(href = "previous", rels = listOf("self")))
+        )
+        publication.setSelfLink("http://manifest.json")
+
+        assertEquals(
+            "http://manifest.json",
+            publication.linkWithRel("self")?.href
+        )
+    }
+
+    @Test fun `get {baseUrl} computes the URL from the {self} link`() {
+        val publication = createPublication(
+            links = listOf(Link(href = "http://domain.com/path/manifest.json", rels = listOf("self")))
+        )
+        assertEquals(
+            URL("http://domain.com/path/"),
+            publication.baseUrl
+        )
+    }
+
+    @Test fun `get {baseUrl} when missing`() {
+        assertNull(createPublication().baseUrl)
     }
 
     @Test fun `find the first {Link} matching the given predicate`() {
@@ -162,28 +427,6 @@ class PublicationTest {
         )
 
         assertNull(publication.coverLink)
-    }
-
-    @Test fun `set {self} link`() {
-        val publication = createPublication()
-        publication.setSelfLink("http://manifest.json")
-
-        assertEquals(
-            "http://manifest.json",
-            publication.linkWithRel("self")?.href
-        )
-    }
-
-    @Test fun `set {self} link replaces existing {self} link`() {
-        val publication = createPublication(
-            links = listOf(Link(href = "previous", rels = listOf("self")))
-        )
-        publication.setSelfLink("http://manifest.json")
-
-        assertEquals(
-            "http://manifest.json",
-            publication.linkWithRel("self")?.href
-        )
     }
